% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LehmannTypeIINormal.R
\name{dL2norm}
\alias{dL2norm}
\title{The Lehmann Type II Normal Distribution}
\source{
Silva, K.T. A. (2021) Distribuicoes Logistica Exponencial-Inversa de Lehmann Tipo I e II.
}
\usage{
dl2norm(x, xi = 0, eta = 1, alpha, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{xi}{vector of location parameters.}

\item{eta}{vector of scale parameters.}

\item{alpha}{vector of power parameters.}
}
\description{
Density, distribution function, quantile function and random
generation for the Lehmann Type II Normal distribution with location
parameter equal to xi, scale parameter equal to eta and power parameter
equal to alpha.
}
\author{
Silva, K.T.A.; MONTALVO, G.S.A.
}
