% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LehmannType1LIE.R
\encoding{UTF-8}
\name{LehmannType1LIE}
\alias{LehmannType1LIE}
\alias{dllie}
\alias{pllie}
\title{The Lehmann Type I Logistic Inverse Exponential}
\usage{
dllie(x, beta, lambda, alpha)
pllie(x, beta, lambda, alpha)
}
\arguments{
\item{x}{vector of quantiles.}

\item{beta}{vector of positive shape parameter.}

\item{lambda}{vector of positive scale parameter.}

\item{alpha}{vector of positive power parameters.}
}
\value{
dllie gives the density, pllie gives the distribution function,
qllie gives the quantile function, and rllie generates random sample.
}
\description{
Density, distribution function, quantile function and random
generation for the Lehmann Type I Logistic Inverse Exponential distribution with location
parameter equal to xi, scale parameter equal to eta and power parameter
equal to alpha.
}
\references{
Silva, Keyliane Travassos Almeida da. Distribuições Logística
Exponencial-Inversa de Lehmann Tipo I e II.–2021. 46 f. Trabalho de
Conclusão de Curso (graduação) – Universidade Federal do Ceará, Centro de
Ciências, Curso de Estatística, Fortaleza, 2021. Orientação:
Prof. Dr. Gualberto Segundo Agamez Montalvo.
}
\author{
Silva, K.T.A.; MONTALVO, G.S.A.
}
