% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LehmannType2LIE.R
\encoding{UTF-8}
\name{LehmannType2LIE}
\alias{LehmannType2LIE}
\alias{dl2lie}
\alias{pl2lie}
\title{The Lehmann Type II Logistic Inverse Exponential}
\usage{
dl2lie(x, beta, lambda, alpha)
pl2lie(x, beta, lambda, alpha)
}
\arguments{
\item{x}{vector of quantiles.}

\item{beta}{vector of positive shape parameter.}

\item{lambda}{vector of positive scale parameter.}

\item{alpha}{vector of positive power parameters.}
}
\value{
dl2lie gives the density, pl2lie gives the distribution function,
ql2lie gives the quantile function, and rl2lie generates random sample.
}
\description{
Density, distribution function, quantile function and random
generation for the Lehmann Type II Logistic Exponential Inverse distribution with location
parameter equal to xi, scale parameter equal to eta and power parameter
equal to alpha.
}
\references{
Silva, Keyliane Travassos Almeida da. Distribuições Logística
Exponencial-Inversa de Lehmann Tipo I e II.–2021. 46 f. Trabalho de
Conclusão de Curso (graduação) – Universidade Federal do Ceará, Centro de
Ciências, Curso de Estatística, Fortaleza, 2021. Orientação:
Prof. Dr. Gualberto Segundo Agamez Montalvo.
}
\author{
Silva, K.T.A.; MONTALVO, G.S.A.
}
