% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LehmannTypeIINormal.R
\encoding{UTF-8}
\name{LehmannTypeIINormal}
\alias{LehmannTypeIINormal}
\alias{dL2norm}
\alias{rL2norm}
\title{The Lehmann Type II Normal Distribution}
\usage{
dL2norm(x, xi = 0, eta = 1, alpha, log = FALSE)
rL2norm(n, xi = 0, eta = 1, alpha)
}
\arguments{
\item{x}{vector of quantiles.}

\item{xi}{vector of location parameters.}

\item{eta}{vector of scale parameters.}

\item{alpha}{vector of power parameters.}

\item{log}{logical; if TRUE, probabilities p are givem log(p)}

\item{n}{number of observations.}
}
\value{
dL2norm gives the density, pL2norm gives the distribution function,
qL2norm gives the quantile function, and rL2norm generates random sample.
}
\description{
Density, distribution function, quantile function and random
generation for the Lehmann Type II Normal distribution with location
parameter equal to xi, scale parameter equal to eta and power parameter
equal to alpha.
}
\references{
Silva, Keyliane Travassos Almeida da. Distribuições Logística
Exponencial-Inversa de Lehmann Tipo I e II.–2021. 46 f. Trabalho de
Conclusão de Curso (graduação) – Universidade Federal do Ceará, Centro de
Ciências, Curso de Estatística, Fortaleza, 2021. Orientação:
Prof. Dr. Gualberto Segundo Agamez Montalvo.
}
\author{
Silva, K.T.A.; MONTALVO, G.S.A.
}
